+++ web/src/routes/+layout.svelte
@@ -1,57 +1,61 @@
 <script context="module" lang="ts">
-  import { createApiClient, loadFetch } from '$lib/api';
-  import type { Load } from '@sveltejs/kit';
-  import initWasm from '$lib/wasm';
-  import { browser } from '$app/env';
+  throw new Error("@migration task: Check code was safely removed (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292722)");
 
-  export const load: Load = async function load({ fetch }) {
-    if (browser) {
-      await initWasm();
-    }
-    fetch = loadFetch(fetch);
-    let [inputList, actionList, actionCategoryList, executorList, accountTypeList, accountList]: [
-      Input[],
-      Action[],
-      ActionCategory[],
-      ExecutorInfo[],
-      AccountType[],
-      AccountPublicInfo[]
-    ] = await Promise.all([
-      fetch('/api/inputs').then((r) => r.json()),
-      fetch('/api/actions').then((r) => r.json()),
-      fetch('/api/action_categories').then((r) => r.json()),
-      fetch('/api/executors').then((r) => r.json()),
-      fetch('/api/account_types').then((r) => r.json()),
-      fetch('/api/accounts').then((r) => r.json()),
-    ]);
+  // import { createApiClient, loadFetch } from '$lib/api';
+  // import type { Load } from '@sveltejs/kit';
+  // import initWasm from '$lib/wasm';
+  // import { browser } from '$app/env';
 
-    let inputs = new Map(inputList.map((i) => [i.input_id, i]));
-    let actions = new Map(actionList.map((a) => [a.action_id, a]));
-    let actionCategories = new Map(actionCategoryList.map((a) => [a.action_category_id, a]));
-    let executors = new Map(executorList.map((e) => [e.name, e]));
-    let accountTypes = new Map(accountTypeList.map((a) => [a.account_type_id, a]));
-    let accounts = new Map(accountList.map((a) => [a.account_id, a]));
+  // export const load: Load = async function load({ fetch }) {
+  //   if (browser) {
+  //     await initWasm();
+  //   }
+  //   fetch = loadFetch(fetch);
+  //   let [inputList, actionList, actionCategoryList, executorList, accountTypeList, accountList]: [
+  //     Input[],
+  //     Action[],
+  //     ActionCategory[],
+  //     ExecutorInfo[],
+  //     AccountType[],
+  //     AccountPublicInfo[]
+  //   ] = await Promise.all([
+  //     fetch('/api/inputs').then((r) => r.json()),
+  //     fetch('/api/actions').then((r) => r.json()),
+  //     fetch('/api/action_categories').then((r) => r.json()),
+  //     fetch('/api/executors').then((r) => r.json()),
+  //     fetch('/api/account_types').then((r) => r.json()),
+  //     fetch('/api/accounts').then((r) => r.json()),
+  //   ]);
 
-    return {
-      props: {
-        inputs,
-        actions,
-        actionCategories,
-        accountTypes,
-        accounts,
-        executors,
-      },
-      stuff: {
-        inputs,
-        actions,
-        actionCategories,
-        executors,
-      },
-    };
-  };
+  //   let inputs = new Map(inputList.map((i) => [i.input_id, i]));
+  //   let actions = new Map(actionList.map((a) => [a.action_id, a]));
+  //   let actionCategories = new Map(actionCategoryList.map((a) => [a.action_category_id, a]));
+  //   let executors = new Map(executorList.map((e) => [e.name, e]));
+  //   let accountTypes = new Map(accountTypeList.map((a) => [a.account_type_id, a]));
+  //   let accounts = new Map(accountList.map((a) => [a.account_id, a]));
+
+  //   return {
+  //     props: {
+  //       inputs,
+  //       actions,
+  //       actionCategories,
+  //       accountTypes,
+  //       accounts,
+  //       executors,
+  //     },
+  //     stuff: {
+  //       inputs,
+  //       actions,
+  //       actionCategories,
+  //       executors,
+  //     },
+  //   };
+  // };
 </script>
 
 <script lang="ts">
+  throw new Error("@migration task: Add data prop (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292707)");
+
   import '../app.css';
   import { page } from '$app/stores';
   import { createDarkStore, cssDarkModePreference } from '$lib/styles';
