+++ web/src/routes/actions/+page.svelte
@@ -3,7 +3,7 @@
   import { getHeaderTextStore } from '$lib/header';
   import Card from '$lib/components/Card.svelte';
   import Labelled from '$lib/components/Labelled.svelte';
-  import ActionEditor from './_ActionEditor.svelte';
+  import ActionEditor from '../_ActionEditor.svelte';
   import Modal, { ModalOpener } from '$lib/components/Modal.svelte';
   import { Action } from '$lib/api_types';
   import clone from 'just-clone';
